monitors_as_code/requirements.txt
pandas==2.1.1
pyyaml==6.0.1
